file structure refactored
