+++ src/main.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import App from "./App.jsx";
-import { Home } from "./pages/Home/Home";
+import { Home } from "./pages/Home";
 import { Auth_Layout } from "./pages/Auth/Auth_Layout";
 import { Error } from "./pages/Error/Error";
 import "./index.css";
